requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
httpx==0.27.0
jinja2==3.1.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.2.2
tiktoken==0.7.0
# Optional: openai client (remove if using another LLM provider)
openai==1.44.0

notion-client==2.2.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
pydub==0.25.1
